requirements.txt
alembic==1.4.3
appdirs==1.4.4
bleach==3.2.1
blinker==1.4
certifi==2020.11.8
click==7.1.2
distlib==0.3.1
dnspython==2.0.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Dropzone==1.5.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-PageDown==0.3.0
Flask-Reuploaded==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
ordereddict==1.1
packaging==20.4
Pillow==8.0.1
pipenv==2020.11.15
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
virtualenv==20.1.0
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==0.16.0
WTForms==2.3.3
